Fix bug when -c option is passed
When a coin with two or more words as name is passed when the -c option
is enabled, the application only took the first word as coin name (in fact, is commander.js who defines this behaviour), producing an error on the request. The solution implemented consist in adding a "-" instead of " " when the name is passed.